refactor: move SendMqttMessageAwaitReply to comms/utils/
